requirements.txt
fastapi
uvicorn
google-generativeai
requests
beautifulsoup4
Pillow
dotenv
torch
python-dotenv
openai
pillow
flask
flask_cors